foodfix/requirements.txt
django
mysqlclient
PyMySQL
djangorestframework
django-filter
